Added a new Helper.cs function that changes color on text depending on their current status called WriteWithStatusColor ().
I also made a new method in JobManager.cs that looks through unanswered old jobs of 14 days and shows them to the user and similarly a
ShowNonRejectJobsTwoWeeksMax() that filters through rejected and job applications older than 2 weeks old and displays them to the user as a bonus.
Readjusted the mock data to with with the requirments for the sake of presentation on run time.

Generally cleaned up the code a bit and some more bugs like forgetting to add .FirstOrDefault() when chaning job status in the UpdateStatus() block of code.

Aside from some minor debugging and clean-up. That should be it!